+++ loraham_daemon/radio_controller.h
@@ -29,6 +29,8 @@ struct RadioController {
     std::atomic<bool> tx_busy;
     std::atomic<bool> cad_active;
     std::atomic<bool> getrssi_active;
+    std::atomic<bool> tx_result_active;
+    uint16_t tx_result_seq;
 
     DaemonRadioStats stats;
 
@@ -62,6 +64,8 @@ static inline void radio_controller_init(RadioController<RadioT> *ctrl,
     ctrl->tx_busy.store(false);
     ctrl->cad_active.store(false);
     ctrl->getrssi_active.store(false);
+    ctrl->tx_result_active.store(false);
+    ctrl->tx_result_seq = 0;
 
     daemon_radio_stats_init(&ctrl->stats);